# Conflicts:
#	identity_data/build.gradle.kts
#	identity_data/src/androidUnitTest/kotlin/net/thechance/mena/identity/data/repository/AddressesRepositoryImplTest.kt
#	identity_data/src/androidUnitTest/kotlin/net/thechance/mena/identity/data/repository/UserRepositoryImplTest.kt
#	identity_presentation/build.gradle.kts

+++ trends_presentation/src/commonMain/kotlin/net/thechance/mena/trends/presentation/screen/user_reel/UserReelScreen.kt
@@ -1,8 +1,10 @@
 package net.thechance.mena.trends.presentation.screen.user_reel
 
 import androidx.compose.animation.animateContentSize
+import androidx.compose.foundation.background
 import androidx.compose.foundation.border
 import androidx.compose.foundation.clickable
+import androidx.compose.foundation.interaction.MutableInteractionSource
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
@@ -13,20 +15,28 @@ import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.size
+import androidx.compose.foundation.pager.VerticalPager
+import androidx.compose.foundation.pager.rememberPagerState
 import androidx.compose.foundation.shape.CircleShape
 import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.getValue
+import androidx.compose.runtime.remember
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.draw.clip
+import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.graphics.painter.Painter
 import androidx.compose.ui.layout.ContentScale
+import androidx.compose.ui.semantics.Role
 import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.unit.dp
+import androidx.compose.ui.zIndex
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
+import app.cash.paging.compose.collectAsLazyPagingItems
 import coil3.compose.rememberAsyncImagePainter
 import mena.trends_presentation.generated.resources.Res
+import mena.trends_presentation.generated.resources.avatar_image
 import mena.trends_presentation.generated.resources.confirmation_message
 import mena.trends_presentation.generated.resources.delete
 import mena.trends_presentation.generated.resources.delete_reel
@@ -45,14 +55,14 @@ import net.thechance.mena.designsystem.presentation.component.icon.Icon
 import net.thechance.mena.designsystem.presentation.component.image.Image
 import net.thechance.mena.designsystem.presentation.component.scaffold.Scaffold
 import net.thechance.mena.designsystem.presentation.component.text.Text
-import net.thechance.mena.designsystem.presentation.theme.theme.MenaTheme
 import net.thechance.mena.designsystem.presentation.theme.theme.Theme
 import net.thechance.mena.trends.presentation.navigation.LocalNavController
+import net.thechance.mena.trends.presentation.navigation.Route
 import net.thechance.mena.trends.presentation.shared.util.ObserveAsEffect
 import net.thechance.mena.trends.presentation.shared.util.gradientShadow
+import net.thechance.mena.trends.presentation.video_player.VideoPlayer
 import org.jetbrains.compose.resources.painterResource
 import org.jetbrains.compose.resources.stringResource
-import org.jetbrains.compose.ui.tooling.preview.Preview
 import org.koin.compose.viewmodel.koinViewModel
 
 @Composable
@@ -65,6 +75,7 @@ internal fun UserReelScreen(
     ObserveAsEffect(viewModel.effect) { effect ->
         when (effect) {
             UserReelEffect.NavigateBack -> navController.popBackStack()
+            UserReelEffect.NavigateToPublisherProfile -> navController.navigate(Route.ManageReels)
         }
     }
 
@@ -79,22 +90,37 @@ private fun UserReelScreenContent(
     state: UserReelState,
     listener: UserReelInteractionListener
 ) {
+
     Scaffold(
         overlays = {
             dialog(isVisible = state.isConfirmationDialogVisible) {
                 Dialog(
                     title = stringResource(Res.string.delete_reel),
                     message = stringResource(Res.string.confirmation_message),
-                    buttonText = stringResource(Res.string.delete),
-                    dismissOnBackPress = true,
-                    dismissOnClickOutside = true,
                     isVisible = state.isConfirmationDialogVisible,
                     onDismiss = { listener.onDismissConfirmationDialog() },
-                    onActionClick = { listener.onConfirmDeleteClick() },
+                    dismissOnBackPress = true,
+                    dismissOnClickOutside = true,
                     onCancelClick = { listener.onDismissConfirmationDialog() },
-                    dialogCornerShape = RoundedCornerShape(12.dp),
+                    dialogCornerShape = RoundedCornerShape(Theme.radius.md),
                     cancelBackgroundShape = RoundedCornerShape(50),
-                    contentPadding = PaddingValues(16.dp)
+                    contentPadding = PaddingValues(Theme.spacing._16),
+                    actionButtons = {
+                        Text(
+                            text = stringResource(Res.string.delete),
+                            color = Theme.colorScheme.error,
+                            style = Theme.typography.label.medium,
+                            modifier = Modifier
+                                .align(Alignment.End)
+                                .padding(top = 24.dp, bottom = Theme.spacing._12, end = Theme.spacing._8)
+                                .clickable(
+                                    onClick = { listener.onConfirmDeleteClick() },
+                                    role = Role.Button,
+                                    indication = null,
+                                    interactionSource = remember { MutableInteractionSource() }
+                                )
+                        )
+                    }
                 )
             }
 
@@ -102,21 +128,21 @@ private fun UserReelScreenContent(
                 Dialog(
                     title = stringResource(Res.string.success_delete_title),
                     message = stringResource(Res.string.success_delete_message),
-                    buttonText = "",
-                    dismissOnBackPress = true,
-                    dismissOnClickOutside = true,
                     isVisible = state.isReelDeleted == true && state.error == null,
                     onDismiss = {
                         listener.onDismissSuccessDialog()
                         listener.onBackClick()
                     },
+                    dismissOnBackPress = true,
+                    dismissOnClickOutside = true,
                     onCancelClick = {
                         listener.onDismissSuccessDialog()
                         listener.onBackClick()
                     },
-                    dialogCornerShape = RoundedCornerShape(12.dp),
+                    dialogCornerShape = RoundedCornerShape(Theme.radius.md),
                     cancelBackgroundShape = RoundedCornerShape(50),
-                    contentPadding = PaddingValues(16.dp)
+                    contentPadding = PaddingValues(Theme.spacing._16),
+                    actionButtons = {}
                 )
             }
 
@@ -124,43 +150,46 @@ private fun UserReelScreenContent(
                 Dialog(
                     title = stringResource(Res.string.fail_delete_title),
                     message = stringResource(Res.string.fail_delete_message),
-                    buttonText = "",
-                    dismissOnBackPress = true,
-                    dismissOnClickOutside = true,
                     isVisible = state.error != null,
                     onDismiss = { listener.onDismissErrorDialog() },
+                    dismissOnBackPress = true,
+                    dismissOnClickOutside = true,
                     onCancelClick = { listener.onDismissErrorDialog() },
-                    dialogCornerShape = RoundedCornerShape(12.dp),
+                    dialogCornerShape = RoundedCornerShape(Theme.radius.md),
                     cancelBackgroundShape = RoundedCornerShape(50),
-                    contentPadding = PaddingValues(16.dp)
+                    contentPadding = PaddingValues(Theme.spacing._16),
+                    actionButtons = {}
                 )
             }
-        },
-        topBar = { TopAppBar(onBackClick = listener::onBackClick) }
+        }
     ) {
-        Box(modifier = Modifier.fillMaxSize()) {
-            UsersReAct(
-                viewCount = state.viewsCount.toString(),
-                likeCount = state.likesCount.toString(),
-                onDeleteClick = listener::onDeleteClick,
-                modifier = Modifier
-                    .align(Alignment.BottomEnd)
-                    .padding(end = Theme.spacing._16, bottom = 140.dp)
-            )
 
-            PublisherDetails(
-                userName = state.username,
-                timeOfPublish = state.createdAt,
-                description = state.description,
-                avatar = state.thumbnail,
-                modifier = Modifier.align(Alignment.BottomCenter),
-                isDescriptionExpanded = state.isDescriptionExpanded,
-                onDescriptionClick = listener::onDescriptionClick
-            )
+        val reels = state.reels.collectAsLazyPagingItems()
+        val pagerState = rememberPagerState(
+            initialPage = 0,
+            pageCount = { reels.itemCount },
+        )
 
-            Box(
-                modifier = Modifier.fillMaxWidth().height(height = 118.dp).gradientShadow()
-            )
+        TopAppBar(onBackClick = listener::onBackClick, modifier = Modifier.zIndex(5f))
+
+        VerticalPager(
+            state = pagerState,
+            modifier = Modifier.fillMaxSize(),
+            key = { page -> reels[page]?.id ?: page },
+        ) { page ->
+
+            reels[page]?.let { reel ->
+                ReelContent(
+                    reel = reel,
+                    shouldRender = (pagerState.currentPage == page),
+                    isDescriptionExpanded = state.isDescriptionExpanded,
+                    onDeleteClick = listener::onDeleteClick,
+                    onDescriptionClick = listener::onDescriptionClick,
+                    onPublisherInfoClick = listener::onPublisherInfoClick,
+                    incrementViewsCount = { listener.increaseReelView(reel.id) },
+                    onLikeClick = { listener.onLikeClick(reel.id) }
+                )
+            }
         }
     }
 }
@@ -175,7 +204,8 @@ private fun TopAppBar(
         modifier = modifier
             .fillMaxWidth()
             .gradientShadow()
-            .padding(horizontal = Theme.spacing._16).padding(top = 8.dp),
+            .padding(horizontal = Theme.spacing._16)
+            .padding(top = Theme.spacing._8),
         contentPadding = PaddingValues(0.dp),
         leadingContent = {
             Icon(
@@ -188,16 +218,61 @@ private fun TopAppBar(
 }
 
 @Composable
-private fun PublisherDetails(
+private fun ReelContent(
+    reel: UserReelUiState,
+    shouldRender: Boolean,
+    isDescriptionExpanded: Boolean,
+    onDeleteClick: () -> Unit,
+    onDescriptionClick: (isCollapsed: Boolean) -> Unit,
+    onPublisherInfoClick: () -> Unit,
+    incrementViewsCount: () -> Unit,
+    onLikeClick: () -> Unit,
+) {
+    VideoPlayer(
+        modifier = Modifier.background(Color.Black),
+        url = reel.videoUrl,
+        isReelVisible = shouldRender,
+        onVideoPlaying = incrementViewsCount
+    ) {
+        Box(Modifier.fillMaxSize()) {
+            UsersReAct(
+                viewCount = reel.viewsCount.toString(),
+                likeCount = reel.likesCount.toString(),
+                isCurrentUserOwner = reel.isCurrentUserOwner,
+                onDeleteClick = onDeleteClick,
+                onLikeClick = onLikeClick,
+                isLiked = reel.isLiked,
+                modifier = Modifier.align(Alignment.BottomEnd)
+                    .padding(end = Theme.spacing._16, bottom = 140.dp)
+            )
+
+            PublisherInfo(
+                userName = reel.username,
+                timeOfPublish = reel.createdAt.toString(),
+                description = reel.description,
+                avatar = reel.profileImageUrl,
+                modifier = Modifier.align(Alignment.BottomCenter),
+                isDescriptionExpanded = isDescriptionExpanded,
+                onDescriptionClick = onDescriptionClick,
+                onPublisherInfoClick = onPublisherInfoClick.takeIf { reel.isCurrentUserOwner } ?: {}
+            )
+
+            Box(modifier = Modifier.fillMaxWidth().height(height = 118.dp).gradientShadow())
+        }
+    }
+}
+
+@Composable
+private fun PublisherInfo(
     avatar: String,
     userName: String,
     timeOfPublish: String,
     isDescriptionExpanded: Boolean,
-    onDescriptionClick: (Boolean) -> Unit,
+    onPublisherInfoClick: () -> Unit,
+    onDescriptionClick: (isCollapsed: Boolean) -> Unit,
     description: String,
     modifier: Modifier = Modifier
 ) {
-
     Column(
         modifier = modifier
             .fillMaxWidth()
@@ -206,16 +281,16 @@ private fun PublisherDetails(
     ) {
         Row(
             horizontalArrangement = Arrangement.spacedBy(space = Theme.spacing._8),
-            modifier = Modifier
-                .padding(top = Theme.spacing._8)
+            modifier = Modifier.padding(top = Theme.spacing._8)
         ) {
 
             Image(
                 painter = rememberAsyncImagePainter(avatar),
                 modifier = Modifier.size(size = 40.dp).clip(shape = CircleShape)
-                    .border(shape = CircleShape, width = 0.5.dp, color = Theme.colorScheme.stroke),
+                    .border(shape = CircleShape, width = 0.5.dp, color = Theme.colorScheme.stroke)
+                    .clickable { onPublisherInfoClick() },
                 contentScale = ContentScale.Crop,
-                contentDescription = "avatar image"
+                contentDescription = stringResource(Res.string.avatar_image)
             )
 
             Column(Modifier.padding(bottom = Theme.spacing._16)) {
@@ -224,6 +299,7 @@ private fun PublisherDetails(
                     color = Theme.colorScheme.primary.onPrimary,
                     style = Theme.typography.label.medium,
                     modifier = Modifier.padding(vertical = Theme.spacing._2)
+                        .clickable { onPublisherInfoClick() },
                 )
 
                 Text(
@@ -252,7 +328,10 @@ private fun PublisherDetails(
 private fun UsersReAct(
     likeCount: String,
     viewCount: String,
+    isCurrentUserOwner: Boolean,
+    isLiked: Boolean,
     onDeleteClick: () -> Unit,
+    onLikeClick: () -> Unit,
     modifier: Modifier = Modifier
 ) {
     Column(
@@ -261,20 +340,24 @@ private fun UsersReAct(
     ) {
         ReActIcon(
             icon = painterResource(resource = Res.drawable.ic_like),
+            onClick = onLikeClick,
+            tint = if (isLiked) Color.White else Theme.colorScheme.shadeTertiary,
             label = likeCount
         )
 
         ReActIcon(
             icon = painterResource(resource = Res.drawable.ic_eye),
+            isClickEnabled = false,
             label = viewCount
         )
 
-        ReActIcon(
-            icon = painterResource(resource = Res.drawable.ic_delete),
-            label = stringResource(Res.string.delete),
-            onClick = { onDeleteClick() },
-
+        if (isCurrentUserOwner) {
+            ReActIcon(
+                icon = painterResource(resource = Res.drawable.ic_delete),
+                label = stringResource(Res.string.delete),
+                onClick = { onDeleteClick() },
             )
+        }
     }
 }
 
@@ -283,6 +366,8 @@ private fun ReActIcon(
     icon: Painter,
     label: String,
     modifier: Modifier = Modifier,
+    isClickEnabled: Boolean = true,
+    tint: Color = Theme.colorScheme.shadeTertiary,
     onClick: () -> Unit = {}
 ) {
     Column(modifier, horizontalAlignment = Alignment.CenterHorizontally) {
@@ -290,9 +375,9 @@ private fun ReActIcon(
             painter = icon,
             contentDescription = stringResource(Res.string.react),
             modifier = Modifier
-                .padding(bottom = 8.dp)
-                .clickable { onClick() },
-            tint = Theme.colorScheme.shadeTertiary
+                .padding(bottom = Theme.spacing._8)
+                .clickable(enabled = isClickEnabled) { onClick() },
+            tint = tint
         )
         Text(
             text = label,
@@ -300,13 +385,5 @@ private fun ReActIcon(
             color = Theme.colorScheme.shadeTertiary,
             textAlign = TextAlign.Center
         )
-    }
-}
-
-@Preview
-@Composable
-private fun Preview() {
-    MenaTheme {
-        UserReelScreen()
     }
 }
